AUTONEWART-05 Header module
Added adaptive header
