+++ .env
@@ -1,3 +1,3 @@
-REACT_APP_API_LIVESITE = AutoNewArt
-REACT_APP_API_LOGO = http://localhost:3000/logo.png
-REACT_APP_API_URL = http://localhost:3000
+REACT_APP_LIVESITE = AutoNewArt
+REACT_APP_LOGO = http://localhost:3000/logo.png
+REACT_APP_URL = http://localhost:3000
