+++ src/components/DrawerMenu/index.ts
@@ -1,0 +1,3 @@
+import DrawerMenu from "./DrawerMenu";
+
+export default DrawerMenu;
