+++ src/components/DrawerMenu/interfaces.ts
@@ -1,0 +1,5 @@
+export interface IDrawerMenu {
+  openMenu: boolean;
+  setOpenMenu: (value: boolean) => void;
+  menuBody: string | null;
+}
