+++ src/components/DrawerMenu/styles.ts
@@ -1,23 +1,19 @@
 import { COLOR } from "constants/colors";
 
-export const contactButtonStyles = {
-  color: COLOR.ACCENT,
-  fontWeight: "400",
-  svg: {
-    fill: COLOR.TRETIARY,
-  },
-};
-
-export const menuButtonStyles = {
+export const drawerButtonStyles = {
   flexDirection: "column",
   background: "#5c5c5c",
   border: "none",
   color: COLOR.ACCENT,
   textTransform: "none",
-  padding: "5px 10px",
+  textAlign: "center",
+  padding: "10px",
   fontWeight: "400",
+  lineHeight: "1",
   width: "150px",
-  gap: "5px",
+  minHeight: "150px",
+  height: "-webkit-fill-available",
+  gap: "10px",
   svg: {
     fill: COLOR.TRETIARY,
     transition: "250ms",
