+++ src/components/HeaderContacts/index.ts
@@ -1,0 +1,3 @@
+import HeaderContacts from "./HeaderContacts";
+
+export default HeaderContacts;
