+++ src/components/index.ts
@@ -1,3 +1,6 @@
+export { default as DrawerMenu } from "./DrawerMenu";
+export { default as HeaderContacts } from "./HeaderContacts";
 export { default as HeaderMenu } from "./HeaderMenu";
 export { default as HeadMeta } from "./HeadMeta";
+export { default as Logo } from "./Logo";
 export { default as Slider } from "./Slider";
