+++ src/styles/media.scss
@@ -1,6 +1,12 @@
-$tablet: 990px;
+$desktop: 1200px;
+$tablet: 900px;
 $mobile: 550px;
 
+@mixin desktop {
+  @media (max-width: #{$desktop}) {
+    @content;
+  }
+}
 @mixin tablet {
   @media (max-width: #{$tablet}) {
     @content;
